# api/docs/swagger.yaml
definitions:
...
  models.DoctorModel:
    properties:
      access_token:
        type: string
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      email:
        type: string
      end_work_year:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      password:
        type: string
      phone_number:
        type: string
      salary:
        type: number
      spec_ids:
        items:
          type: integer
        type: array
      start_work_year:
        type: string
      updated_at:
        type: string
      work_years:
        type: integer
    type: object
  models.DoctorReq:
    properties:
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      department_id:
        type: integer
      email:
        type: string
      end_work_year:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      salary:
        type: number
      spec_ids:
        items:
          type: integer
        type: array
      start_work_year:
        type: string
      work_years:
        type: integer
    type: object
  models.DoctorResp:
    properties:
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      email:
        type: string
      end_work_year:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      password:
        type: string
      phone_number:
        type: string
      salary:
        type: number
      spec_ids:
        items:
          type: integer
        type: array
      start_work_year:
        type: string
      updated_at:
        type: string
      work_years:
        type: integer
    type: object
  models.DoctorUpdateReq:
    properties:
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      department_id:
        type: integer
      end_work_year:
        type: string
      full_name:
        type: string
      id:
        type: string
      salary:
        type: number
      start_work_year:
        type: string
      work_years:
        type: integer
    type: object
...
  models.ListDoctors:
    properties:
      count:
        type: integer
      doctors:
        items:
          $ref: '#/definitions/models.DoctorResp'
        type: array
    type: object
...
  models.LoginReqModel:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginRespDoctor:
    properties:
      access_token:
        type: string
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      email:
        type: string
      end_work_year:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      password:
        type: string
      phone_number:
        type: string
      salary:
        type: number
      spec_ids:
        items:
          type: integer
        type: array
      start_work_year:
        type: string
      updated_at:
        type: string
      work_years:
        type: integer
    type: object
...
  models.RegisterRespModel:
    properties:
      message:
        type: string
    type: object
...
  models.Status:
    properties:
      message:
        type: string
    type: object
...
host: localhost:7070
info:
  contact: {}
  description: api-gateway
  title: Clinic system
  version: "1.0"
paths:
...
  /v1/doctor/{id}:
    get:
      consumes:
      - application/json
      description: Get user
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DoctorResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get doctor by id
      tags:
      - Doctor
  /v1/doctor/create:
    post:
      consumes:
      - application/json
      description: Create a new doctor with the provided details
      parameters:
      - description: Create doctor
        in: body
        name: DoctorInfo
        required: true
        schema:
          $ref: '#/definitions/models.DoctorReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DoctorModel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: create doctor
      tags:
      - Doctor
  /v1/doctor/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete doctor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Status'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete doctor
      tags:
      - Doctor
  /v1/doctor/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Login
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.LoginReqModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LoginRespDoctor'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: login doctor
      tags:
      - Doctor
  /v1/doctor/register:
    post:
      consumes:
      - application/json
      description: Register a new doctor with the provided details
      parameters:
      - description: Register doctor
        in: body
        name: DoctorInfo
        required: true
        schema:
          $ref: '#/definitions/models.DoctorReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RegisterRespModel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: register doctor
      tags:
      - Doctor
  /v1/doctor/update/{id}:
    put:
      consumes:
      - application/json
      description: Update doctor
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: Update Doctor
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/models.DoctorUpdateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DoctorResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: update doctor
      tags:
      - Doctor
  /v1/doctor/verify/{email}/{code}:
    get:
      consumes:
      - application/json
      description: Verify a doctor with code sent to their email
      parameters:
      - description: email
        in: path
        name: email
        required: true
        type: string
      - description: code
        in: path
        name: code
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Status'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: verify doctor
      tags:
      - Doctor
  /v1/doctors/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: get all doctors
      parameters:
      - description: page
        in: path
        name: page
        type: string
      - description: limit
        in: path
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ListDoctors'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get doctors' list
      tags:
      - Doctor
  /v1/doctors/{page}/{limit}/{department_id}:
    get:
      consumes:
      - application/json
      description: get all doctors by department id
      parameters:
      - description: page
        in: path
        name: page
        type: string
      - description: limit
        in: path
        name: limit
        type: string
      - description: department_id
        in: path
        name: department_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ListDoctors'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get doctors' list by department id
      tags:
      - Doctor
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"